github: lint pull-requests with golangci-lint
The new lint job runs in parallel with tests and provides line-level PR
annotations automatically.

We lint the entire module instead of just the changes files (with the
only-new-issues option) because it evaluates into flags of golangci-lint
that we know have problems. This has the side-effect of users having to
fix problems in code they did not write when linters evolve implicitly.

Smart caching is handled automatically by the golangci-lint action to
minimize CI runtime.